requirements.txt
Flask==2.3.3
python-whois==0.8.0
requests==2.31.0
tld==0.13
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
reportlab==4.0.4